build(hack): require derive feature on dev tests
Since mongod_derive::Mongo relies on mongod_derive::Bson, we have no
clean way of exposing this due to the mongo re-export. So for now lets
just make sure that this feature is enabled during dev. Unfortunately
this means users will always have to use the feature derive if they want
to use the derive macros. The split crate approach will not work.