feat: convert collection to use serde
With BSON & MongoDB 2.0 which is clearly geared at using serde for
conversion using the interim document type for Collection no longer
makes sense. For that reason Collection has been changed to require
serde, this should not be too big an issue as customer derives can be
used to obtain the old behaviour. This change also allows for Cursor to
return concrete types addressing #2.